pdr-rpa-approve: service.module.spec to test automatic config retrieval
